Move pipeline exchange from CTE to pipeline level
