+++ src/execution/operator/scan/physical_column_data_scan.cpp
@@ -112,7 +112,7 @@ void PhysicalColumnDataScan::BuildPipelines(Pipeline &current, MetaPipeline &met
 			D_ASSERT(cte_sink->type == PhysicalOperatorType::CTE);
 			auto &cte = cte_sink->Cast<PhysicalCTE>();
 			auto &source = cte_source->Cast<PhysicalCTEConsumerSource>();
-			if (cte.TryRegisterFanoutPipeline(current, source.consumer_idx)) {
+			if (cte.TryRegisterDirectConsumer(current, source.consumer_idx)) {
 				auto current_pipeline = current.shared_from_this();
 				current.SetExternalInput();
 				current.AddExternalFinishDependency(cte_dependency);
